Bump to version 1.0 Now uses getopt for parameter parsing, allowing long form usage and more descriptive help output. Added -f --foreground option. Fixed unsigned/signed type comparison warning.

+++ arp.c
@@ -71,13 +71,13 @@ int arp_recv(int sock, ether_arp_frame *frame)
 
 	nread=recv(sock, &packet, sizeof(packet), 0);
 
-	if (nread > sizeof(ether_arp_frame)) {
+	if (nread > (int)sizeof(ether_arp_frame)) {
 		nread=sizeof(ether_arp_frame);
 	}
 
 	memcpy(frame, &packet, nread);
 
-	if (nread < sizeof(ether_arp_frame)) {
+	if (nread < (int)sizeof(ether_arp_frame)) {
 		memset(((char*)frame) + nread, 0, sizeof(ether_arp_frame) - nread);
 	}
